Completed the Creature Tracker
In addition to adding the species to each player's tracker, it also
adds the unique id of a species seen. There are functions to check if
more than 3 of the same species is seen. In addition there is an added
method for checking if a specific id has been seen.